view.pyのルーティング機能をBlueprintで分離
・view.pyにひとまとめにしていたルーティング機能をBlueprintで分離
・blueprints/auth, blueprints/profile, blueprints/studyにroute.pyを作成し、ルーティング機能を分離
・__init__.pyのcreate_appにblueprintをインポート、登録
