+++ src/blueprints/auth/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+auth_bp = Blueprint('auth', __name__)
+
+from . import routes
