+++ src/blueprints/study/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+study_bp = Blueprint('study', __name__)
+
+from . import routes
