+++ src/models/__init__.py
@@ -1,0 +1,6 @@
+from .users import User
+from .fields import Field
+from .studyLogs import StudyLog
+
+__all__ = ['User', 'Field', 'StudyLog']
+
