+++ src/templates/auth/login.html
@@ -9,7 +9,7 @@
     <div class="auth-content">
       <h1>ログイン</h1>
       <div class="auth-form">
-        <form action="{{ url_for('login_process') }}" method="post">
+        <form action="{{ url_for('auth.login_process') }}" method="post">
           <div class="auth-input">
             <input
               placeholder="メールアドレス"
@@ -30,8 +30,8 @@
         </form>
       </div>
       <div class="auth-link">
-        <a href="{{ url_for('signup_view') }}">新規登録はこちら</a>
-        <a href="{{ url_for('password_reset_view') }}"
+        <a href="{{ url_for('auth.signup_view') }}">新規登録はこちら</a>
+        <a href="{{ url_for('auth.password_reset_view') }}"
           >パスワードをお忘れの方はこちら</a
         >
       </div>
