+++ src/templates/auth/password_reset.html
@@ -10,7 +10,7 @@ content %}
     <div class="auth-content">
       <h1>パスワード再設定</h1>
       <div class="auth-form">
-        <form action="{{ url_for('password_reset_process') }}" method="post">
+        <form action="{{ url_for('auth.password_reset_process') }}" method="post">
           <div class="auth-input">
             <input
               placeholder="メールアドレス"
@@ -37,8 +37,8 @@ content %}
         </form>
       </div>
       <div class="auth-link">
-        <a href="{{ url_for('login_view') }}">ログイン画面はこちら</a>
-        <a href="{{ url_for('signup_view') }}">新規登録はこちら</a>
+        <a href="{{ url_for('auth.login_view') }}">ログイン画面はこちら</a>
+        <a href="{{ url_for('auth.signup_view') }}">新規登録はこちら</a>
       </div>
     </div>
   </div>
