+++ src/templates/profile/password_update.html
@@ -8,11 +8,11 @@
 <div class="main-container">
   <div class="auth-section">
     <div class="auth-href">
-      <a href="{{ url_for('profile_edit_view', user_id=current_user.user_id) }}"
+      <a href="{{ url_for('profile.profile_edit_view', user_id=current_user.user_id) }}"
         >プロフィール</a
       >
       <a
-        href="{{ url_for('password_update_view', user_id=current_user.user_id) }}"
+        href="{{ url_for('profile.password_update_view', user_id=current_user.user_id) }}"
         >パスワード変更</a
       >
     </div>
@@ -22,7 +22,7 @@
       <h1>パスワード変更</h1>
       <div class="auth-form">
         <form
-          action="{{ url_for('password_update_process', user_id=current_user.user_id) }}"
+          action="{{ url_for('profile.password_update_process', user_id=current_user.user_id) }}"
           method="post"
         >
           <div class="auth-input">
