+++ src/templates/study/index.html
@@ -11,7 +11,7 @@
       <h1>{{ current_user.username }}さんの学習履歴</h1>
     </div>
     <form
-      action="{{ url_for('get_graph_stats', user_id=current_user.user_id) }}"
+      action="{{ url_for('study.get_graph_stats', user_id=current_user.user_id) }}"
       method="post"
       id="graphForm"
       class="home-display-wrap"
