+++ src/templates/study/study_logs_list.html
@@ -21,7 +21,7 @@
       </div>
       <div class="study-date">
         <form
-          action="{{ url_for('study_logs_list_process', user_id=current_user.user_id) }}"
+          action="{{ url_for('study.study_logs_list_process', user_id=current_user.user_id) }}"
           method="post"
           id="study-date-form"
         >
